requirements.txt
apturl==0.5.2
attrs==20.2.0
Automat==20.2.0
awscli==1.18.142
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
boto==2.49.0
boto3==1.15.1
botocore==1.18.1
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
chompjs==1.0.16
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==5.0.0
constantly==15.1.0
crayons==0.4.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipython==7.13.0
ipython-genutils==0.2.0
itemadapter==0.1.0
itemloaders==1.0.3
jedi==0.15.2
jmespath==0.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==1.4.4
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parsel==1.6.0
parso==0.5.2
pexpect==4.6.0
phantomjs==1.3.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyDispatcher==2.0.5
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.5.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
Scrapy==2.3.0
SecretStorage==2.3.1
selenium==3.141.0
Selenium-Screenshot==1.7.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.0.1
systemd-python==234
traitlets==4.3.3
Twisted==20.3.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
w3lib==1.22.0
wadllib==1.3.3
wcwidth==0.1.8
webdriver-manager==3.2.2
webencodings==0.5.1
xkit==0.0.0
zope.interface==5.1.0

Flask~=1.1.2
psutil~=5.7.2